requirements.txt
# Outbound Prospecting Pipeline - Dependencies
requests
google-api-python-client
google-auth-oauthlib
